rename: change use auth hook path
