+++ components/GlobalHeader/GlobalHeader.tsx
@@ -6,7 +6,7 @@ import Box from '@mui/material/Box';
 import ButtonBase from '@mui/material/ButtonBase';
 import IconButton from '@mui/material/IconButton';
 import DarkModeSwitch from 'components/DarkModeSwitch';
-import useAuth from 'hooks/context/useAuth';
+import useAuth from 'hooks/auth/useAuth';
 import useColorMode from 'hooks/context/useColorMode';
 import { useRouter } from 'next/router';
 import React, { useState } from 'react';
