+++ hooks/auth/useKakaoLoginRedirect.tsx
@@ -1,5 +1,5 @@
 import { postLoginWithKakao } from 'helpers/axios';
-import useAuth from 'hooks/context/useAuth';
+import useAuth from 'hooks/auth/useAuth';
 import { useRouter } from 'next/router';
 import { ApiLoginKakaoErrorData } from 'pages/api/login/kakao';
 import { useEffect, useRef, useState } from 'react';
